requirements.txt
# VaaniSeva - Python Dependencies
# AI for Bharat Hackathon 2026

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Twilio - Telephony
twilio>=9.0.0

# HTTP
requests>=2.31.0

# Local dev
python-dotenv>=1.0.0

# Document processing (for knowledge-updater Lambda)
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0

# HTTP and API
requests>=2.31.0
urllib3>=2.0.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
pymongo>=4.5.0  # If using MongoDB
redis>=5.0.0    # For caching

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging and Monitoring
python-json-logger>=2.0.7
sentry-sdk>=1.40.0

# Phone Number Handling
phonenumbers>=8.13.0

# Date and Time
python-dateutil>=2.8.2
pytz>=2023.3

# Text Processing
nltk>=3.8.1
langdetect>=1.0.9

# API Framework (if building REST API)
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0

# Data Validation
marshmallow>=3.20.0
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=4.2.0  # AWS mocking
responses>=0.24.0  # HTTP mocking

# Code Quality
black>=23.11.0
flake8>=6.1.0
pylint>=3.0.0
mypy>=1.7.0
isort>=5.12.0

# Development Tools
ipython>=8.17.0
jupyter>=1.0.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0
aioboto3>=12.0.0

# Utilities
python-slugify>=8.0.0
humanize>=4.9.0
tqdm>=4.66.0

# AWS Lambda Specific
mangum>=0.17.0  # For FastAPI on Lambda

# Monitoring and Telemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Rate Limiting
ratelimit>=2.2.1
limits>=3.6.0

# Caching
cachetools>=5.3.0
diskcache>=5.6.0

# Configuration Management
configparser>=6.0.0
argparse>=1.4.0

# Indian Regional Data
indic-nlp-library>=0.81

# WhatsApp/SMS (Optional future expansion)
flask-twilio>=0.0.5

# Background Tasks
celery>=5.3.0  # If using task queues
redis>=5.0.0   # Celery broker

# Database ORM (if needed)
sqlalchemy>=2.0.0
alembic>=1.12.0

# API Documentation
pyyaml>=6.0.1

# Performance
gunicorn>=21.2.0
gevent>=23.9.0

# Compression
zstandard>=0.22.0

# Time Series (for analytics)
influxdb-client>=1.38.0

# Graph/Network Analysis (for call flow)
networkx>=3.2.0

# Cost Tracking
stripe>=7.0.0  # If implementing payments

# Localization
babel>=2.13.0
